Merge FeatureData and DataStore into one
Reduce the number of data files required by having only one source of
data for the mapper and the turn reader. This requires some extra
processing of the turn's text to match with data in the database and
some changing of the data in the database to match with the turn's text.