Calculate size of DiffMapModel based on used area of secondary
Only expand the bound of the DiffMapModel if the secondary map's used
area is outside the primary map's bound. This still doesn't actually
do the right thing as you'll never save anything that is outside the
primary's bound. This is to fix up another day - along with allowing a
map to be resized via the UI. Then the primary can be resized when the
DiffMapModel is constructed and everything with save nice.