Fix bug where note is too long
A user could enter a note that is longer than 255 characters. When this
is written to a JIM data file the length wraps round and prevents the
map file from being opened. Updated notes text field in the UI to stop
taking input after 250 characters and also updated the JIM data factory
to only ever write a maximum of 250 characters into a note entry.